Remove unused `DataInputReadable`
